Moving protobuf in dedicated folder
